Updated style.css
Implemented a new idea I had to make the sidebar collapsible with the idea that It becomes more flexible to different screen sizes. This is my attempt to fix an earlier Mobile issue I had described where the sidebar would cover half the screen, making navigation very difficult. 

I like working in VS Code for any additions or major changes when typing out my HTML/CSS as the VS software assists with spell check and helps me with completing code lines as I am learning HTML and CSS as I go about coding this.  This is my reasoning for having this change be a big copy and paste, as I was working on it incrementally on VS code and testing on the Live Preview editor on there before releasing it to GitHub Sites with the chance of it being possible bugged. 

I plan to go through and comment specific CSS elements to help assist me with picking up on the Syntax and getting familiar with what specific parts of the code do when I work more on this.  I have been doing most of my learning on YouTube and through reading different Internet documentation specific to my fix/desires for the Site.