Updated index.html + JS
It has been a while since adding anything, I was really determined to fix the Mobile issue I had discovered a month or so ago. Through researching when I had time and countless failure to make any change lol, I had decided that making the Sidebar collapsible would make it work the best without hindering the vision I had. I found out through researching on YouTube that anything "Event" related is more of JavaScript and I was far limited in CSS to do what I wanted. 

I did not know any JavaScript until making this change so most of the time between now and the last Commit was due to me studying up and trying to get familiar with JavaScript syntax and the language itself. My syntax is definitely not perfect, and it has room for improvement, but I happily did get the sidebar button to work. This Commit is going to lead to me testing it on Mobile to see the effectiveness. 

I did have to ask GitHub CoPilot how I could add a delay i.e.. setTimeout() which I am not proud of doing. I had run into an issue where the button was not showing any animation when it was clicked, which I learned is because of the duration of the Animation. The reset of the 'Fall" class was happening quicker than the Animation took to finish. I regret having to use Assistance like this, but with my lack of Knowledge and experience in the Language, this was something I did not understand that I had to 'manage' and as such I was unsure of a Fix. Even through using assistance, this taught me for my next implementation of animation to have the delay. 

This site helped me tremendously https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date
for trying to implement a RTC, which I felt comfortable with after having added a button. This article explained the Date() object which I do understand Objects/Classes decently from working with classes C++. It went further down to specific Functions related to find the year, day, month, etc which through looking at some sample HTML code on various websites, I took a go at it after seeing someone elses code on Stack (https://stackoverflow.com/questions/50438363/show-real-time-clock-on-html-using-javascript)
I did not want to copy it at all, so I took from it the general premise and just tested as I went. 
     The setInterval() was commented by someone a little down in response to the original code posted, I added it after seeing it. Some of the aspects of JS are very similar to C++ which allowed me to add code rather easily when I knew what Functions and Syntax I needed to use. 
      I plan to study JavaScript more so I can add some more features to my website and to allow me to create much more dynamic websites in the Future.