New index.html layout
Complete revamp of our main page or index. Changed appearance by color and layout. I didn't like the basic design of the previous, this idea to me looks much cleaner and more professional. I enjoy not having the sidebar as having to manually toggle it was just an inconvenience. 

This commit added new sections, changed class names, added a container and box layout which will be used in the future. Finally, added a JS button function to scroll to a specific section. Plan to add actual content soon and to keep tweaking the layout.