Updated style.css
The guts of the HTML commit. Complete change of colors and look - this builds a top bar instead of the sidebar. I made it so it was at the top using flex for a more modern look and to keep it out of the way at all times. 
Also added a glow effect to our new introduction(scroll) button. The rest of the page is almost the same except now our 'content' section is using the display: flex layout to center the content within the box- equal margins on both sides and the target section of the scroll button. I plan for this to be used as a little introduction section with pictures and text. Currently empty just to create the shape. Moved JS clock to bottom right to not block any important elements. 
Next steps are:
-Get more heavily into design; I want a different image then the Tux image. 
-Add content to the introduction and make the main page generally completed. 
- Add more links and things to show off.