New features and font!
Took away project button and incorporated a grid of my big projects. Added overlay for when user hovers. These will eventually be turned into in browser modals to actually demonstrate the code. Will add projects button back for smaller projects. Making more items the 'monospace' font as I like the code look to it. Finally, working on making the background stretch across the whole screen as I noticed a gap on the right. As such, I removed the horizontal spacebar to try to remove that scroll-ability, no features require scrolling so this for the time being works. I am unsure of mobile compatibility at this time, this push will be the test run so to speak for mobile.