New Additions
Added a custom Windows debloat script through a downloadable link to allow people who don't have GitHub to download and run it. Added boxes/hover changes to link.
Created external JS for security purposes. Due to time constraints, It is not the nicest looking code but I plan to go through and revise- just wanted to incorporate the "big" changes.