Major Update
HUGE additions to the site. This is the guts of it all with I believe 3 new functions since the last commit. Added a typewriter and createart function, these are used with the ASCII art which is also a new addition. I wanted a computer vibe and thought the page printing out ascii art would fit this and it does perfect. Worked with GSAP for the first time to implement a splash screen which I had heavily desired to do something similar to falling rain from The Matrix. 

I was able to accomplish this through many days of learning this new library which I plan to incorporate much more. JavaScript is still a new language to me, forgive any improper code. I would love any feedback if there is any to give!