Fixing ASCII display issues on mobile
One if statement to test the screen width of the device accessing the site. If its smaller than 600px ( can be adjusted) then it simply will not display the art. This was done after testing on my phone and the ASCII art taking forever to spit out due to the typewriter function and it not being able to get as many characters in each row. 

In order to not be unfriendly to mobile, I figured this is not a necessary aspect to the site at all when compared to being device friendly. Need to create if statement to have main-content appear in a shorter duration if screen is smaller (i.e no ASCII art and the timing of display is set based off the art).