Refactor CSS to remove scrollbar
Updated CSS styles to take the unappealing scrollbar off the screen. Scrolling is still enabled/possible but it is more clean and there isn't a chunk of white and gray which I find to make a site go from custom to a professionally done site. Simple couple lines, copied and pasted from my VS Code.